updated parent component
