+++ .gitignore
@@ -42,3 +42,6 @@ yarn-error.log*
 next-env.d.ts
 
 backend/typst-wrapper/typst-wrapper
+
+# local notes file
+commands.md
